create new API for getById

+++ source/modules/articles/controllers/delete.ts
@@ -1,9 +1,25 @@
 import { NextFunction, Response } from "express";
-import { ERRORS } from "../../../config/enums";
+import { ERRORS, USER_TYPES } from "../../../config/enums";
 import { ArticlesTable } from "../model";
 
 export async function deleteArticle(request: any, response: Response, next: NextFunction) {
 	try {
+		const article: any = await ArticlesTable.findOne({
+			where: {
+				articleId: request.params.articleId,
+			},
+		});
+
+		if (!article) {
+			next(new Error(ERRORS.ARTICLE_NOT_FOUND));
+		}
+
+		if (article && request.user.type === USER_TYPES.CONTENT_CREATOR) {
+			if (article.ownerId !== request.user.userId) {
+				next(new Error(ERRORS.ARTICLE_NOT_YOURS));
+			}
+		}
+
 		const rowCount = await ArticlesTable.destroy({
 			where: {
 				articleId: request.params.articleId,
